+++ pravega-sensor-collector/src/test/resources/RawFileIngestService.properties
